backend/requirements.txt
djangorestframework-simplejwt
icalendar
django